refactor: dissolve ParameterShared into cohesive parameter helpers
